modify: 공지사항 조회수, 파일업로드 리셋 기능 구현

+++ src/main/java/com/myrium/domain/NoticeVO.java
@@ -18,8 +18,8 @@ public class NoticeVO {
     private Date createdAt;
     private String updatedBy;
     private Date updatedAt;
-    private Integer readCnt;
-    private Integer hasFiles;
-    private Integer isDeleted;
+    private Long readCnt;
+    private int hasFiles;
+    private int isDeleted;
     private MultipartFile[] files;
 }
